selenium/python/requirements.txt
mailosaur==7.17.1
selenium==4.21.0
python-dotenv==1.0.1